// ui/src/app/LPC/enums/female/LPC-female-special.enum.ts
export enum LPCFemaleSpecialEnum {
  CAPE_BLACK = 'CAPE_BLACK',
  CAPE_BLUE = 'CAPE_BLUE',
  CAPE_BROWN = 'CAPE_BROWN',
  CAPE_GRAY = 'CAPE_GRAY',
  CAPE_GREEN = 'CAPE_GREEN',
  CAPE_LAVENDER = 'CAPE_LAVENDER',
  CAPE_MAROON = 'CAPE_MAROON',
  CAPE_PINK = 'CAPE_PINK',
  CAPE_RED = 'CAPE_RED',
  CAPE_WHITE = 'CAPE_WHITE',
  CAPE_YELLOW = 'CAPE_YELLOW'
}